requirements.txt
seaborn
imblearn
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
tensorflow-macos==2.13.0
matplotlib==3.7.1
